+++ lib/thumbnail_slider.dart
@@ -1,2 +1,3 @@
+export 'src/duration_extension.dart';
 export 'src/seek_bar.dart';
 export 'src/thumbnail_value_indicator.dart';